CRUD operations for core StorySumthings and Folders are working. Now to just make the UI...

+++ elements/Element.cs
@@ -1,7 +1,7 @@
 namespace sumstories.elements;
 
-public abstract class Element(int ID) {
-	public readonly int ID = ID;
+public abstract class Element(long ID) {
+	public readonly long ID = ID;
 	public string Name { get; set; } = "New Element";
 
 	public Category Category { get; set; } = Category.NONE;
